Create [Example] Crash Reporting Functionality in O3DE.md
This is an example commit that creates a pull request to sandbox how we might share a document in the form of a markdown file with partners so that external teams can comment and make changes. The body text would be replaced with the content of the RFC document and discussion could happen on the pull request directly.